cron_job/requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.20.0
apscheduler==3.9.1
requests==2.28.0
pika==1.2.1